requirements.txt
# Media File Organizer Dependencies
requests>=2.28.0
python-dotenv>=1.0.0
PyYAML>=6.0
